@@ -4,49 +4,54 @@ import "errors"
 
 var i any = -1
 
+// Iterator is a channel-based iterator for traversing the linked list.
 type Iterator[T any] <-chan T
 
+// ListNode represents a node in a doubly linked list.
 type ListNode[T comparable] struct {
 	val        T
 	next, prev *ListNode[T]
 }
 
+// NewListNode creates a new node with the given value.
 func NewListNode[T comparable](val T) *ListNode[T] {
 	return &ListNode[T]{val: val}
 }
 
+// DoublyLinkedList represents a generic doubly linked list.
 type DoublyLinkedList[T comparable] struct {
 	size       int
 	head, tail *ListNode[T]
 }
 
+// NewLinkedList initializes and returns a new empty doubly linked list.
 func NewLinkedList[T comparable]() *DoublyLinkedList[T] {
 	return &DoublyLinkedList[T]{size: 0}
 }
 
-// Clear - empty this linked list
+// Clear removes all elements from the list, resetting it to empty.
 func (dl *DoublyLinkedList[T]) Clear() {
 	dl.head = nil
 	dl.tail = nil
 	dl.size = 0
 }
 
-// Size - return the size of the linked list, O(1)
+// Size returns the number of elements in the list. O(1)
 func (dl *DoublyLinkedList[T]) Size() int {
 	return dl.size
 }
 
-// IsEmpty - return true if linked list is empty false otherwise, O(1)
+// IsEmpty checks if the linked list is empty. O(1)
 func (dl *DoublyLinkedList[T]) IsEmpty() bool {
 	return dl.size == 0
 }
 
-// Add - add element to the tail of the linked list, O(1)
+// Add appends an element to the end of the list. O(1)
 func (dl *DoublyLinkedList[T]) Add(elem T) (bool, error) {
 	return dl.AddLast(elem)
 }
 
-// AddLast - add a node to the tail of the linked list, O(1)
+// AddLast appends a new element at the tail of the list. O(1)
 func (dl *DoublyLinkedList[T]) AddLast(elem T) (bool, error) {
 	newNode := NewListNode(elem)
 	if dl.IsEmpty() {
@@ -61,7 +66,7 @@ func (dl *DoublyLinkedList[T]) AddLast(elem T) (bool, error) {
 	return true, nil
 }
 
-// AddFirst - add an element to the beginning of the linked list, O(1)
+// AddFirst inserts a new element at the head of the list. O(1)
 func (dl *DoublyLinkedList[T]) AddFirst(elem T) (bool, error) {
 	newNode := NewListNode(elem)
 	if dl.IsEmpty() {
@@ -76,9 +81,9 @@ func (dl *DoublyLinkedList[T]) AddFirst(elem T) (bool, error) {
 	return true, nil
 }
 
-// AddAt - add at specific index, O(n)
+// AddAt inserts an element at a specific index in the list. O(n)
 func (dl *DoublyLinkedList[T]) AddAt(idx int, elem T) (bool, error) {
-	if idx < 0 || idx > dl.size {
+	if idx < 0 || idx >= dl.size {
 		return false, errors.New("invalid index")
 	}
 	newNode := NewListNode(elem)
@@ -109,7 +114,7 @@ func (dl *DoublyLinkedList[T]) AddAt(idx int, elem T) (bool, error) {
 	return true, nil
 }
 
-// PeekFirst - check the value of the first node if it exists, O(1)
+// PeekFirst returns the value of the first element. O(1)
 func (dl *DoublyLinkedList[T]) PeekFirst() (T, error) {
 	if dl.IsEmpty() {
 		return i.(T), errors.New("linked list empty")
@@ -117,7 +122,7 @@ func (dl *DoublyLinkedList[T]) PeekFirst() (T, error) {
 	return dl.head.val, nil
 }
 
-// PeekLast - check the value of the last node if it exists, O(1)
+// PeekLast returns the value of the last element. O(1)
 func (dl *DoublyLinkedList[T]) PeekLast() (T, error) {
 	if dl.IsEmpty() {
 		return i.(T), errors.New("linked list empty")
@@ -125,7 +130,7 @@ func (dl *DoublyLinkedList[T]) PeekLast() (T, error) {
 	return dl.tail.val, nil
 }
 
-// RemoveFirst - remove the first value at the head of the linked list, O(1)
+// RemoveFirst removes and returns the first element. O(1)
 func (dl *DoublyLinkedList[T]) RemoveFirst() (T, error) {
 	if dl.IsEmpty() {
 		return i.(T), errors.New("linked list empty")
@@ -141,7 +146,7 @@ func (dl *DoublyLinkedList[T]) RemoveFirst() (T, error) {
 	return value, nil
 }
 
-// RemoveLast - remove the last value at the tail of the linked list, O(1)
+// RemoveLast removes and returns the last element. O(1)
 func (dl *DoublyLinkedList[T]) RemoveLast() (T, error) {
 	if dl.IsEmpty() {
 		return i.(T), errors.New("linked list empty")
@@ -157,7 +162,7 @@ func (dl *DoublyLinkedList[T]) RemoveLast() (T, error) {
 	return value, nil
 }
 
-// Remove - remove a node with provided value, O(N)
+// Remove deletes the first occurrence of a given element. O(n)
 func (dl *DoublyLinkedList[T]) Remove(elem T) (T, error) {
 	if dl.IsEmpty() {
 		return i.(T), errors.New("linked list empty")
@@ -189,7 +194,7 @@ func (dl *DoublyLinkedList[T]) Remove(elem T) (T, error) {
 	}
 }
 
-// RemoveAt - remove node at idx, O(N)
+// RemoveAt removes and returns the element at a specific index. O(n)
 func (dl *DoublyLinkedList[T]) RemoveAt(idx int) (T, error) {
 	if idx < 0 || idx >= dl.size {
 		return i.(T), errors.New("invalid index")
@@ -218,7 +223,7 @@ func (dl *DoublyLinkedList[T]) RemoveAt(idx int) (T, error) {
 	return i.(T), errors.New("invalid index")
 }
 
-// find the index of the element in linked list, O(N)
+// indexOf finds the index of an element in the list. O(n)
 func (dl *DoublyLinkedList[T]) indexOf(elem T) (int, error) {
 	if dl.IsEmpty() {
 		return -1, errors.New("linked list empty")
@@ -236,7 +241,7 @@ func (dl *DoublyLinkedList[T]) indexOf(elem T) (int, error) {
 	return -1, errors.New("element not found in linked list")
 }
 
-// Contains - check if an element exists in the list, O(N)
+// Contains checks if an element exists in the list. O(n)
 func (dl *DoublyLinkedList[T]) Contains(elem T) (bool, error) {
 	result, err := dl.indexOf(elem)
 	if err != nil {
@@ -245,6 +250,7 @@ func (dl *DoublyLinkedList[T]) Contains(elem T) (bool, error) {
 	return result >= 0, nil
 }
 
+// Iterate returns a channel-based iterator for traversing the list.
 func (dl *DoublyLinkedList[T]) Iterate() Iterator[T] {
 	iterChan := make(chan T)
 	go func() {
